Fix bug 1
This commit fixes a bug where some items were not being added to orders because they were created before the order was created. The fix involves updating the logic to check for items created after the order was created.

Fix bug 2

This commit also fixes a bug where items created by an order did not have the proper tags assigned to them. We have now updated the code to add tags based on the order information.

New feature

In addition to these bug fixes, we have also added a new order number field to the order model to make it easier to track orders. We have also updated the code to save order emails to a utils file with a specific naming convention for easier identification.